predictive-maintenance/data_loader/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
kafka-python>=2.0.2
